requirements.txt
python-telegram-bot==13.0.0
Flask==2.3.3
pytz==2023.3
requests==2.31.0
dateparser==1.1.8
gunicorn==21.2.0
Werkzeug
asyncio==3.4.3